:art: redesigned comments section

+++ portfolio-frontend/src/components/EditDeleteProjectButtons.tsx
@@ -1,21 +1,27 @@
-import { DeleteIcon, EditIcon } from '@chakra-ui/icons';
-import { Box, IconButton, Link } from '@chakra-ui/react';
+import { AddIcon, DeleteIcon, EditIcon, ExternalLinkIcon, HamburgerIcon, RepeatIcon } from '@chakra-ui/icons';
+import { AlertDialog,Text, AlertDialogBody, AlertDialogContent, AlertDialogFooter, AlertDialogHeader, AlertDialogOverlay, Box, Button, Flex, IconButton, Link, Menu, MenuButton,Circle, MenuItem, MenuList } from '@chakra-ui/react';
 import React from 'react'
 import NextLink from "next/link"
+import {MdMoreHoriz, MdDelete, MdModeEdit, MdMoreVert} from "react-icons/md"
 import { useDeleteProjectMutation, useMeQuery } from '../generated/graphql';
 
 interface EditDeleteProjectButtonsProps {
     id: number
     creatorId: number
+    name: string
 
 }
 
 export const EditDeleteProjectButtons: React.FC<EditDeleteProjectButtonsProps> = ({
     id,
-    creatorId
+    creatorId,
+    name
 }) => {
     const [, deleteProject] = useDeleteProjectMutation()
     const [{data: meData}] = useMeQuery()
+    const [isOpen, setIsOpen] = React.useState(false)
+    const onClose = () => setIsOpen(false)
+    const cancelRef = React.useRef()
     
     if (meData?.me?.id !== creatorId) {
         return null
@@ -23,23 +29,76 @@ export const EditDeleteProjectButtons: React.FC<EditDeleteProjectButtonsProps> =
 
  
         return (
+            <>
+            <Menu>
+              
+              <MenuButton
+                as={IconButton}
+                aria-label="Options"
+                icon={<MdMoreVert />}
+                _hover={{bg:"none"}}
+                _focus={{bg:"none"}}
+                variant="unstyled"
+              />
+             
+              <MenuList>
+              <NextLink href="/project/edit/[id]" as={`/project/edit/${id}`} >
+                <MenuItem icon={<MdModeEdit />}>
+                  Edit 
+                </MenuItem>
+                </NextLink>
+                <MenuItem  color="red" icon={<MdDelete />} onClick={() => setIsOpen(true)}>
+                  Delete 
+                </MenuItem>
+             
+              </MenuList>
+            </Menu>
             
-            <Box>
-               
-                <IconButton 
-                marginRight={4}
+        <AlertDialog
+          isOpen={isOpen}
+          leastDestructiveRef={cancelRef}
+          onClose={onClose}
+        >
+          <AlertDialogOverlay>
+            <AlertDialogContent>
+              <AlertDialogHeader fontSize="lg" fontWeight="bold">
+                Delete {name}
+              </AlertDialogHeader>
+  
+              <AlertDialogBody>
+                  <Flex>
+                      <Text>Are you sure you want to delete</Text>
+                      <Text marginLeft="1" fontWeight="bold" >{name}</Text>
+                      <Text>?</Text>
+                  </Flex>
+                
+              </AlertDialogBody>
+  
+              <AlertDialogFooter>
+                <Button 
+                ref={cancelRef} 
+                onClick={onClose}
+                _hover={{bg:"none"}}
+                _focus={{bg:"none"}}
+                variant="link"
+                >
+                  Cancel
+                </Button>
+                <Button 
+                variant="link"
+                colorScheme="red" 
+                _hover={{bg:"none"}}
+                _focus={{bg:"none"}}
                 onClick={()=> {
                     deleteProject({id})
-                    }}
-                aria-label="Delete project" 
-                icon={<DeleteIcon/>} 
-                />
-            <NextLink href="/project/edit/[id]" as={`/project/edit/${id}`} >
-                <IconButton 
-                as={Link}
-                aria-label="Update project" 
-                icon={<EditIcon/>} />
-            </NextLink>
-        </Box>
+                }} ml={3}>
+                  Delete
+                </Button>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialogOverlay>
+        </AlertDialog>
+
+        </>
         );
 }
